% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bls_cps_extraction.R
\name{cps_employment_data_extraction}
\alias{cps_employment_data_extraction}
\title{CPS emlpoyment data extraction}
\usage{
cps_employment_data_extraction(years, directory)
}
\arguments{
\item{years}{Array containing the year or years for which we want to extract the data}

\item{directory}{Directory where files will be saved}
}
\description{
This function downloads the Occupations files from the BLS Current Population Survey for years 2011 onwards \url{https://www.bls.gov/cps/lfcharacteristics.htm#occind}
All files will be saved in the cps_all_files and mapped to TNR soc taxonomy.
}
\examples{
Extract 2015, 2016 and 2017 CPS data sets from BLS website
years<-c(2015:2017)
directory<-"/data/Stata/us_data/"
cps_employment_data_extraction(years,directory)
}
\keyword{bls}
