% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bls_oes_extraction.R
\name{oes_employment_data_extraction}
\alias{oes_employment_data_extraction}
\title{OES emlpoyment data extraction}
\usage{
oes_employment_data_extraction(years, directory)
}
\arguments{
\item{years}{Array containing the year or years for which we want to extract the data}
}
\description{
This function downloads the "All data" Occupation Employment Survey data from BLS for years 2012 onwards \url{https://www.bls.gov/oes/tables.htm}
All files will be saved in the oes_all_files.
}
\examples{
Extract 2015, 2016 and 2017 "All data" OES data sets from BLS website
years<-c(15:17)
directory<-"/data/Stata/us_data/"
bls_emlpoyment_data_extraction(years,directory)
}
\keyword{bls}
